backend/src/requirements.txt
certifi==2023.7.22
Flask==3.0.0
Flask-Cors==4.0.0
Jinja2==3.1.2
PyJWT==2.8.0
pymongo==4.6.0
python-dotenv==1.0.0
requests==2.31.0